+++ .github/workflows/script/check-js.sh
@@ -7,7 +7,7 @@ if [ ! -z "$(git status --porcelain)" ]; then
     >&2 echo "Failed: Repo should be clean before testing!"
     exit 1
 fi
-# Wipe the lib directory incase there are extra unnecessary files in there
+# Wipe the lib directory in case there are extra unnecessary files in there
 rm -rf lib
 # Generate the JavaScript files
 npm run-script build