+++ src/config-utils.test.ts
@@ -7,6 +7,7 @@ import * as yaml from "js-yaml";
 import * as sinon from "sinon";
 
 import * as api from "./api-client";
+import { CachingKind } from "./caching-utils";
 import {
   CodeQL,
   getCachedCodeQL,
@@ -52,6 +53,7 @@ function createTestInitConfigInputs(
       configInput: undefined,
       buildModeInput: undefined,
       trapCachingEnabled: false,
+      dependencyCachingEnabled: CachingKind.None,
       debugMode: false,
       debugArtifactName: "",
       debugDatabaseName: "",
@@ -347,6 +349,7 @@ test("load non-empty input", async (t) => {
       augmentationProperties: configUtils.defaultAugmentationProperties,
       trapCaches: {},
       trapCacheDownloadTime: 0,
+      dependencyCachingEnabled: CachingKind.None,
     };
 
     const languagesInput = "javascript";